move files to folders with affiliated css file, move css from notes/techs into proper file
