+++ src/components/NavBar/NavBar.css
@@ -1,0 +1,10 @@
+.pages{
+  margin: 0;
+  height: 4.5vh;
+}
+
+button {
+  margin-top: 2px;
+  height: 100%;
+  width: 50%;
+}
