+++ src/components/Title/Title.css
@@ -1,8 +1,3 @@
-.pages{
-  margin: 0;
-  height: 5vh;
-}
-
 h1 {
   margin: 0;
   text-align: center;
