--- src/components/TopNav.css
@@ -1,4 +1,0 @@
-/* .pages{
-  margin: 0;
-  height: 5vh;
-} */
