+++ src/pages/Notes/Notes.css
@@ -9,14 +9,6 @@
   height: 3vh;
 } */
 
-button {
-  /* border: none;
-  background: none; */
-  margin-top: 2px;
-  height: 100%;
-  width: 50%;
-}
-
 /* width */
 ::-webkit-scrollbar {
   width: 10px;
