+++ src/pages/Techs/Techs.css
@@ -1,11 +1,3 @@
-button {
-  /* border: none;
-  background: none; */
-  margin-top: 2px;
-  height: 100%;
-  width: 50%;
-}
-
 /* width */
 ::-webkit-scrollbar {
   width: 10px;
